feat: multi-line node hl support
rust doc comments do not work.
the way they get parsed is insane.

`/// This is a doc comment`, for example, yields the following nodes:
comment { 0, 0, 0, 25 }
comment.documentation { 0, 0, 0, 25 }
operator { 0, 2, 0, 3 }
comment { 0, 2, 0, 3 }

why are the last two even there at all?
it already knows it's a comment and it still parses a single '/'
as both a comment _and_ an operator. insane